Tidy up by splitting API traits from internal implementation
