+++ bpa/src/dispatcher/mod.rs
@@ -14,7 +14,7 @@ mod restart;
 pub(crate) struct Dispatcher {
     tasks: hardy_async::TaskPool,
     processing_pool: hardy_async::BoundedTaskPool,
-    store: Arc<storage::Store>,
+    store: Arc<storage::store::Store>,
     rib: Arc<rib::Rib>,
     keys_registry: Arc<keys::registry::Registry>,
     filter_engine: Arc<filter::FilterEngine>,
@@ -36,7 +36,7 @@ impl Dispatcher {
         poll_channel_depth: core::num::NonZeroUsize,
         processing_pool_size: core::num::NonZeroUsize,
         node_ids: Arc<node_ids::NodeIds>,
-        store: Arc<storage::Store>,
+        store: Arc<storage::store::Store>,
         rib: Arc<rib::Rib>,
         keys_registry: Arc<keys::registry::Registry>,
         filter_engine: Arc<filter::FilterEngine>,
@@ -61,7 +61,7 @@ impl Dispatcher {
         poll_channel_depth: core::num::NonZeroUsize,
         processing_pool_size: core::num::NonZeroUsize,
         node_ids: Arc<node_ids::NodeIds>,
-        store: Arc<storage::Store>,
+        store: Arc<storage::store::Store>,
         rib: Arc<rib::Rib>,
         keys_registry: Arc<keys::registry::Registry>,
         filter_engine: Arc<filter::FilterEngine>,
